Separate test class for ConnectorOffsetBackingStore + one single Future object to be returned to the caller
